third_party/blink/web_tests/flag-specific/highdpi/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/inline-size-containment-vertical-rl-expected.txt

This is a testharness.js-based test.
FAIL inline-size containment only assert_equals: expected 50 but got 20
Harness: the test ran to completion.
